Rename InputSidebar to ToolSidebar
