--- frontend/src/components/InputSidebar/index.ts
@@ -1,3 +1,0 @@
-// components/InputSidebar/index.ts
-export { InputSidebar } from './InputSidebar';
-export type { InputSidebarProps } from './InputSidebar';
