+++ frontend/src/components/ToolSidebar/ToolSelector.tsx
