+++ frontend/src/components/ToolSidebar/index.ts
@@ -1,0 +1,3 @@
+// components/ToolSidebar/index.ts
+export { ToolSidebar } from './ToolSidebar';
+export type { ToolSidebarProps } from './ToolSidebar';
