+++ frontend/src/components/index.ts
@@ -1,5 +1,5 @@
 // components/index.ts
 export * from './Sidebar';
 export * from './ChatInterface';
-export * from './InputSidebar';
+export * from './ToolSidebar';
 export * from './Auth';
